Add basic task loading and running features and rename 'Bask'
A 'Baskfile' is expected to be a file with a series of bash functions
that can be run using this task-running command. Upon running, the
program looks for a file with the name 'Baskfile' in the current
directory and, if one is not found, it continues searching parent
directories until either one is found or the root level of the directory
structure is reached.

Additionally, an `init` function creates a new 'Baskfile' in the current
directory. This file contains a single example function.