Don't run main automatically when the program is sourced elsewhere
In order to make it possible to source the script in other script,
omit the call to _main unless the basename of the caller script matches
the basename of this script.

Because the program is structured such that new functions can be added
at the end of the script, the function that loads these additional
command functions is the `_main` function, which needs to be run last. By
omitting this when the program is sourced, the sourcing program can
consist of a list of functions, followed by the expected `_main` call,
which then loads those functions.