Declare and assign separately to avoid masking return values.
Shellcheck SC2155

In the original code, the return value of mycmd is ignored, and export
will instead always return true. This may prevent conditionals, set -e
and traps from working correctly.

When first marked for export and assigned separately, the return value
of the assignment will be that of mycmd. This avoids the problem.

https://github.com/koalaman/shellcheck/wiki/SC2155