Escape function names to permit "-" and ":" in help
Bash allows ":" and "-" in function names, but they are not allowed
in identifiers for variables. In order to store the description in a
variable that matches the function name, this function escapes the
function name in order to make it suitable for use in a variable
identifier.

As implemented, the function name should transparently escape colons
and hyphens without additional changes to the API.