Add support for definable program help headers.
Help headers are displayed at the beginning of program help and can be
used to provide a basic description for a program. Using an identifiar
that is unlikely to be used for a function name makes it possible to use
the existing `desc` function.

Important note: the ascii art needed extra escaping due to this change.