Add optional command groups in `commands` output.
Grouped command output provides a clearer indication of which
commands are program commands and which ones are defined within the
current scope.

In order to minimize the work done loading the command group arrays,
the builtin command list is hard-coded into a program-wide variable
which is then checked against during command loading. These variables
are later used in the `commands` command to display the commands in
each group.

Some speculative usability concerns are addressed in the display of the
output. Rather than use an excessively-generally name for the commands
defined in the current scope, the name is changed depending on whether
the current scope's commands are defined in a script sourcing this
program or in a context-specific file, like a Baskfile. Additionally,
the local scope's command group  name is not displayed when the list is
empty.

More comments than usual have been added in order to reduce ambiguity
about the potentially-confusing conditional statements.