Remove top comment in generated Baskfiles
The comment at the top of these generated files is causing filetype
settings to not take effect. This comment is arguably redundant since
the usage information for this function is included in the `desc` block.
For now, removing this comment streamlines these generated files and
avoids the syntax / filetype detection issues.