Refactor generated file templates with heredocs.
`printf` statements require listing all of the variables at the end in
the order they occur within the template, which is difficult to keep
track with high numbers of variables. heredocs make it possible to place
the variables directly in the template. This also has the side effect of
making it clearer where the templates begin and end, and it makes it
possible to avoid additional noise from escaping double quotes.

The `<<TEMPLATE || true` form is needed in order to avoid exiting the
program when the heredoc exits with a non-zero status. For more
information, see:
  https://github.com/alphabetum/bash-boilerplate#set--o-errexit--set--e