Use `read -r` to avoid unexpected backslash escaping.
`read -r` causes the backslash to not act as an escape character.
Without the `-r` option, backslashes need to be escaped multiple times,
which is particularly jarring with `HELP_HEADER` descriptions containing
ascii art. This change breaks compatibility with any existing
`HELP_HEADER` in bask scripts, so they will need to be updated.